requirements_1.txt
absl-py==0.10.0
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
bleach==3.2.1
blessings==1.7
breezy==3.1.0
Brlapi==0.7.0
btrfsutil==1.2.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
configobj==5.0.6
contextlib2==0.6.0.post1
cryptography==3.1.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
docutils==0.16
dulwich==0.20.6
entrypoints==0.3
fastimport==0.9.8
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.30.2
h5py==2.10.0
html5lib==1.1
Hydra==2.5
hydra-core==0.11.3
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==1.6.1
include-server==3.3.3
iotop==0.6
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.1.0
ipywidgets==7.5.1
isort==5.6.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.5.1
lensfun==0.3.95
lockfile==0.12.2
louis==3.15.0
lxml==4.5.2
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
meson==0.55.3
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
notebook==6.1.4
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.2.1
ordered-set==4.0.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
patiencediff==0.2.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
powerline-status==2.8.1
progress==1.5
progressbar2==3.53.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.6.0
pwquality==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycairo==1.19.1
pycocotools==2.0
pycodestyle==2.6.0
pycparser==2.20
pycuda==2019.1.2
pycups==2.0.1
pycurl==7.43.0.6
pyelftools==0.26
pyflakes==2.2.0
Pygments==2.7.1
PyGObject==3.38.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.4.0
python-xlib==0.28
pytools==2020.4
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
resolvelib==0.4.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
team==1.0
tensorboard==2.3.0
tensorboard-plugin-wit @ file:///build/python-tensorboard_plugin_wit/src/tensorboard_plugin_wit-1.7.0/wit-pip/release/dist/tensorboard_plugin_wit-1.7.0-py3-none-any.whl
tensorflow-estimator @ file:///build/python-tensorflow-estimator/src/estimator_pip/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
tensorflow-gpu @ file:///build/tensorflow/src/tmpoptcuda/tensorflow_gpu-2.3.1-cp38-cp38-linux_x86_64.whl
tensorflow-serving-api @ file:///tmp/pip/tensorflow_serving_api-2.3.0-py2.py3-none-any.whl
tensorflow-serving-api-gpu @ file:///tmp/pip/tensorflow_serving_api_gpu-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.10.1
toml==0.10.1
torch==1.6.0
torchsummary==1.5.1
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.24
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.3.1
